Update domain status without WlsDomainConfig
