+++ integration-tests/src/test/java/oracle/weblogic/kubernetes/assertions/TestAssertions.java
@@ -12,6 +12,7 @@ import java.util.concurrent.Callable;
 import io.kubernetes.client.openapi.ApiException;
 import io.kubernetes.client.openapi.models.V1Pod;
 import io.kubernetes.client.openapi.models.V1Secret;
+import io.kubernetes.client.util.Yaml;
 import oracle.weblogic.domain.DomainCondition;
 import oracle.weblogic.kubernetes.actions.impl.LoggingExporter;
 import oracle.weblogic.kubernetes.assertions.impl.Apache;
@@ -464,6 +465,7 @@ public class TestAssertions {
 
       if (domain != null && domain.getStatus() != null) {
         List<DomainCondition> domainConditionList = domain.getStatus().getConditions();
+        logger.info(Yaml.dump(domainConditionList));
         for (DomainCondition domainCondition : domainConditionList) {
           if (domainCondition.getType().equalsIgnoreCase(conditionType)) {
             return true;
@@ -520,6 +522,7 @@ public class TestAssertions {
 
       if (domain != null && domain.getStatus() != null) {
         List<DomainCondition> domainConditionList = domain.getStatus().getConditions();
+        logger.info(Yaml.dump(domainConditionList));
         for (DomainCondition domainCondition : domainConditionList) {
           if (domainCondition.getType().equalsIgnoreCase(conditionType)
               && domainCondition.getStatus().equalsIgnoreCase(expectedStatus)) {
